Redo commands
This breaks up the "cron" command into a bunch of commands rather than having
it do faux subcommands. In doing that, I shuffled a bunch of code and tests
around, too.
